+++ src/CMakeLists.txt
@@ -5,8 +5,10 @@ file(GLOB hayai_headers
   hayai_console.hpp
   hayai_console_outputter.hpp
   hayai_default_test_factory.hpp
+  hayai_file_outputter.hpp
   hayai_fixture.hpp
   hayai_outputter.hpp
+  hayai_stream_outputter.hpp
   hayai_test.hpp
   hayai_test_descriptor.hpp
   hayai_test_factory.hpp
